+++ Logging/RockLib.Logging.Analyzers.Test/Verifiers/CSharpAnalyzerVerifier`1.cs
@@ -3,6 +3,7 @@ using Microsoft.CodeAnalysis.CSharp.Testing;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing;
 using Microsoft.CodeAnalysis.Testing.Verifiers;
+using Microsoft.CodeAnalysis.Text;
 using System.Collections.Immutable;
 using System.Threading;
 using System.Threading.Tasks;
@@ -34,6 +35,31 @@ namespace RockLib.Logging.Analyzers.Test
                     .AddPackages(ImmutableArray.Create(
                         new PackageIdentity("RockLib.Logging", "3.0.5")))
             };
+
+            test.ExpectedDiagnostics.AddRange(expected);
+            await test.RunAsync(CancellationToken.None);
+        }
+
+        /// <inheritdoc cref="AnalyzerVerifier{TAnalyzer, TTest, TVerifier}.VerifyAnalyzerAsync(string, DiagnosticResult[])"/>
+        public static async Task VerifyAnalyzerAsync(string source, (string filename, string content) additionalFile, params DiagnosticResult[] expected)
+        {
+            var test = new Test
+            {
+                TestCode = source,
+                ReferenceAssemblies = ReferenceAssemblies.Default
+                    .AddPackages(ImmutableArray.Create(
+                        new PackageIdentity("RockLib.Logging", "3.0.5")))
+            };
+
+            if (additionalFile.filename != null && additionalFile.content != null)
+            {
+                test.SolutionTransforms.Add((solution, projectId) =>
+                {
+                    var documentId = DocumentId.CreateNewId(projectId, debugName: additionalFile.filename);
+                    solution = solution.AddAdditionalDocument(documentId, additionalFile.filename, additionalFile.content);
+                    return solution;
+                });
+            }
 
             test.ExpectedDiagnostics.AddRange(expected);
             await test.RunAsync(CancellationToken.None);