+++ Logging/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+#### Added
+
+- Analyzer and codefix for RockLib0009: Logger has no log providers.
+
 ## 1.0.3 - 2021-08-12
 
 #### Changed